savvystripe/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.2
beautifulsoup4==4.5.3
billiard==3.3.0.23
bs4==0.0.1
carton==0.3
celery==3.1.11
decorator==4.0.11
dj-stripe==0.8.0
Django==1.11
django-allauth==0.31.0
django-appconf==1.0.1
django-bootstrap-form==3.2.1
django-braces==1.11.0
django-cart==1.0.4
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-hstore==1.4.2
django-image-cropping==1.0.4
django-model-utils==2.6.1
django-notifications-hq==1.2
django-recaptcha==1.2.1
django-simple-math-captcha==1.0.7
django-stdimage==2.4.1
django-zebra==0.4.5
docutils==0.13.1
easy-thumbnails==2.3
gevent==1.0.2
greenlet==0.4.10
jsonfield==1.0.3
kombu==3.0.37
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pinax-notifications==4.0.0
pinax-stripe==3.3.0
progressbar2==3.12.0
psycopg2==2.7.1
PyContracts==1.7.6
pyparsing==2.2.0
python-openid==2.2.5
python-utils==2.0.1
pythonwhois==2.4.3
pytz==2016.10
pywhois==0.1
redis==2.9.1
requests-oauthlib==0.8.0
six==1.9.0
snowplow-tracker==0.8.0
stripe==1.48.1
tld==0.7.8
